% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterheatmap.R
\name{clusterheatmap}
\alias{clusterheatmap}
\title{Generates a heat map based on an iClusterVB object}
\usage{
clusterheatmap(fit, nvars = NULL, title = NULL, cols = NULL)
}
\arguments{
\item{fit}{A fitted iClusterVB object.}

\item{nvars}{The number of variables to show; defaults to all.}

\item{title}{Title of the heat map.}

\item{cols}{A vector of colors to use for the clusters; defaults to a random selection of colors.}
}
\value{
Returns a heat map.
}
\description{
Generates a heat map based on an iClusterVB object
}
\examples{
clusterheatmap(fit_iClusterVB, nvars = 50, title = "View 1", cols = c("green","blue","purple", "red"))
}
