% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chmap.R
\name{chmap}
\alias{chmap}
\title{Generates a heat map based on an iClusterVB object}
\usage{
chmap(fit, rho = 0, cols = NULL, title = NULL, ...)
}
\arguments{
\item{fit}{A fitted iClusterVB object.}

\item{rho}{The minimum probability of inclusion for variables shown on the
heatmap. Default is 0. Only useful for VS_method = 1.}

\item{cols}{A vector of colors to use for the clusters. The default is a random selection of colors.}

\item{title}{A character vector or a single value. Title of the heat map. The
default is "View 1 - Distribution 1", ..., "View R - Distribution R".}

\item{...}{Additional arguments to be passed down to
\code{\link[pheatmap]{pheatmap}}}
}
\value{
Returns a heat map for each data view.
}
\description{
Generates a heat map based on an iClusterVB object
}
\examples{
\dontrun{
chmap(fit_iClusterVB, rho = 0.75, title = c("View 1", "View 2"),
cols = c("green", "blue", "purple", "red"))
chmap(fit_iClusterVB)
}
}
