\name{iClusterVB}
\alias{iClusterVB}
\title{A Next-Generation R Package for Bayesian Clustering}
\usage{
iClusterVB(mydata, dist, K = 10, initial_method = "VarSelLCM", VS_method = 0, initial_cluster = NULL, initial_vs_prob = NULL, initial_fit = NULL, initial_omega = NULL, input_hyper_parameters = NULL, max_iter = 200, early_stop = 1, per = 10, convergence_threshold = 1e-4)
}
\description{
iClusterVB—an integrative clustering approach grounded in variational Bayesian inference. iClusterVB allows researchers to navigate high-dimensional datasets comprising mixed-type data, including continuous, categorical, and count variables, as well as multi-view data.
}
\examples{
}
