% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iClusterVB.R
\name{iClusterVB}
\alias{iClusterVB}
\title{Fast Integrative Clustering for High-Dimensional Multi-View Data Using
Variational Bayesian Inference}
\usage{
iClusterVB(
  mydata,
  dist,
  K = 10,
  initial_method = "VarSelLCM",
  VS_method = 0,
  initial_cluster = NULL,
  initial_vs_prob = NULL,
  initial_fit = NULL,
  initial_omega = NULL,
  input_hyper_parameters = NULL,
  max_iter = 200,
  early_stop = 1,
  per = 10,
  convergence_threshold = 1e-04
)
}
\arguments{
\item{mydata}{input data as a list of length R, where R is the number of data
sets.}

\item{dist}{a vector of length R specifying the type of data, or distribution
- dist = 'gaussian' (for continuous data), 'multinomial' (for binary or
categorical data), and 'poisson' (for count data).}

\item{K}{The maximum number of clusters, with a default of 10. The algorithm
will converge to yield a model that is composed of dominant clusters where
redundant clusters are removed, automating the process of determining the
number of clusters.}

\item{initial_method}{the initialization method for the cluster allocation -
"VarSelLCM" for VarSelLCM,  "random" for a random sample, "kproto" for
k-prototypes, "kmeans" for k-means (continuous data only), "mclust" for
mclust(continuous data only), or "lca" for poLCA(categorical data only).
The default method is VarSelLCM.}

\item{VS_method}{the variable selection method -  0 = clustering without
variable selection, 1 = clustering with variable selection. The default is
0, clustering without variable selection.}

\item{initial_cluster}{The initial cluster membership. The default is NULL,
which uses initial_method for initial cluster allocation. If it is not
NULL, it will overwrite the previous initial values setting for this
parameter.}

\item{initial_vs_prob}{the initial variable selection probability, a scalar.
The default is NULL, which assigns a value of 0.5.}

\item{initial_fit}{initial values based on a previously fitted iClusterVB
model (an iClusterVB object). The default is NULL.}

\item{initial_omega}{customized initial values for variable inclusion
probabilities. The default is NULL. If the argument is not NULL, it will
overwrite the previous initial values setting for this parameter. If
VS_method = 1, initial_omega is a list of length R, and each element of the
list is an array with dim=c(N,p[[r]])). N is the sample size and p[[r]] is
the number of variables for dataset r, r = 1,...,R.}

\item{input_hyper_parameters}{the initial hyper-parameters of the prior
distributions for the model. The default is NULL, which assigns alpha_00 =
0.001, mu_00 = 0, s2_00 = 100, a_00 = 1, b_00 = 1,kappa_00 = 1, u_00 = 1,
v_00 = 1.}

\item{max_iter}{the maximum number of iteration of the VB algorithm. The
default is 200.}

\item{early_stop}{whether to stop the algorithm when it converges or continue
until it reaches max_iter. 1 - the algorithm stops when it converges, 0 -
the algorithm stops when it reaches the maximum iteration (regardless of
whether it converges or not). The default is 1, the algorithm stops when it
converges.}

\item{per}{print information every "per" iteration. The default is 10}

\item{convergence_threshold}{define a convergence threshold for the change in
ELBO. The default is 0.0001.}
}
\value{
The iClusterVB function creates an object of class `iClusterVB`, a
  list. The variable inclusion probabilities can be accessed through
  fit$model_parameters$rho[[r]], r = 1,...,R. The clusters can be can be
  accessed through fit$cluster, although a more useful output would be
  table(fit$cluster)|. A summary output can be obtained using the function
  summary.iClusterVB(object, vs_prob,...)
}
\description{
`iClusterVB` offers a novel, fast, and integrative approach to clustering
high-dimensional, mixed-type, and multi-view data, addressing the
computational complexities and challenges of modern biomedical studies. By
employing variational Bayesian inference, iClusterVB facilitates effective
feature selection and identification of disease subtypes, enhancing clinical
decision-making. Leverage iClusterVB to
advance your data analysis and uncover meaningful insights in heterogeneous
disease research.
}
\examples{
iClusterVB(mydata = dat1, dist = c("gaussian", "multinomial", "gaussian", "poisson"), K = 6, initial_method = "kproto")

}
