% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piplot.R
\name{piplot}
\alias{piplot}
\title{Generates a probability inclusion plot based on an iClusterVB object}
\usage{
piplot(
  fit,
  plot_grid = TRUE,
  ylab = "Probability of Inclusion",
  title = NULL,
  ...
)
}
\arguments{
\item{fit}{A fitted iClusterVB object.}

\item{plot_grid}{LOGICAL. Whether to use the \code{\link[cowplot]{plot_grid}}
function from the \bold{cowplot} package. The default is TRUE.}

\item{ylab}{The y-axis label. The default is "Probability of Inclusion".}

\item{title}{The title of the plots. It can be a character vector or a single
value. The default output is "View 1 - Distribution 1", ..., "View R -
Distribution R".}

\item{...}{Additional arguments to add to the
\code{\link[cowplot]{plot_grid}} function.}
}
\value{
Returns a probability inclusion plot or plots.
}
\description{
Generates a probability inclusion plot based on an iClusterVB object
}
\examples{
\dontrun{
piplot(fit_iClusterVB,
  ylab = "Probability", title = NULL, align = "hv",
  nrow = 2, ncol = 2
)
piplot(fit_iClusterVB)
}
}
