% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piplot.R
\name{piplot}
\alias{piplot}
\title{Generates a probability inclusion plot based on an iClusterVB object}
\usage{
piplot(
  fit,
  combined_plot = TRUE,
  ylab = "Probability of Inclusion",
  default_title = TRUE,
  title = NULL
)
}
\arguments{
\item{fit}{A fitted iClusterVB object.}

\item{combined_plot}{LOGICAL. Whether to plot all the plots together or separately; defaults to TRUE (together).}

\item{ylab}{The y-axis label; defaults to "Probability of Inclusion".}

\item{default_title}{LOGICAL; defaults is TRUE, which outputs "View 1,..R - Distribution".}

\item{title}{The title of the plots. It can be a character vector or a single value; default outputs the default title.}
}
\value{
Returns a probability inclusion plot or plots.
}
\description{
Generates a probability inclusion plot based on an iClusterVB object
}
\examples{
piplot(fit_iClusterVB, ylab = "Probability", default_title = TRUE, title = NULL)
piplot(fit_iClusterVB)

}
