% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{plot.iClusterVB}
\alias{plot.iClusterVB}
\title{Generic plot method for `iClusterVB` objects}
\usage{
\method{plot}{iClusterVB}(fit)
}
\arguments{
\item{fit}{A fitted iClusterVB object.}
}
\value{
Returns an ELBO plot, a barplot of cluster proportions, and heatmaps
  for each data view.
}
\description{
Generic plot method for `iClusterVB` objects
}
\examples{
\dontrun{plot(fit_iClusterVB)}
}
