% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{plot.iClusterVB}
\alias{plot.iClusterVB}
\title{Generic plot method for `iClusterVB` objects}
\usage{
\method{plot}{iClusterVB}(x, ...)
}
\arguments{
\item{x}{A fitted iClusterVB object.}

\item{...}{Potential further arguments (unused)}
}
\value{
Returns an evidence lower bound (ELBO) plot and a barplot of cluster
  percentages.
}
\description{
Generic plot method for `iClusterVB` objects
}
\examples{
\dontrun{plot(fit_iClusterVB)}
}
