% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{summary.iClusterVB}
\alias{summary.iClusterVB}
\title{Generic summary method for `iClusterVB` objects}
\usage{
\method{summary}{iClusterVB}(object, rho = 0.5, ...)
}
\arguments{
\item{object}{A fitted iClusterVB object.}

\item{rho}{The minimum posterior inclusion probability of interest to count
the number of variables that >= it. Default is 0.5. Only works for
VS_method = 1.}

\item{...}{Potential further arguments}
}
\value{
Returns a summary list for an `agnes` object.
}
\description{
Generic summary method for `iClusterVB` objects
}
\examples{
summary(fit_iClusterVB)

}
